Fixing somme review comments
